requirements.txt
asyncpraw
python-dotenv
requests
discord.py
